Create models.py
Change some ints to longs.

+++ app/models.py
@@ -204,10 +204,10 @@ class UserStatus(models.Model):
     ectrolium = models.BigIntegerField(default=0, validators = [MinValueValidator(0)])
 
     # Current resource production/decay, calculated in process_tick
-    energy_production = models.IntegerField(default=0)
-    energy_decay = models.IntegerField(default=0)
-    energy_interest = models.IntegerField(default=0) # wookies only
-    energy_income = models.IntegerField(default=0)
+    energy_production = models.BigIntegerField(default=0)
+    energy_decay = models.BigIntegerField(default=0)
+    energy_interest = models.BigIntegerField(default=0) # wookies only
+    energy_income = models.BigIntegerField(default=0)
     mineral_production = models.IntegerField(default=0)
     mineral_interest = models.IntegerField(default=0) # wookies only
     mineral_income = models.IntegerField(default=0)
@@ -221,12 +221,12 @@ class UserStatus(models.Model):
 
     # Misc info that must be recalculated
     num_planets = models.IntegerField(default=1) # number of planets, will get calculated
-    population = models.IntegerField(default=0)
+    population = models.BigIntegerField(default=0)
     networth = models.BigIntegerField(default=1)
-    buildings_upkeep = models.IntegerField(default=0) # stored as a positive number
-    units_upkeep = models.IntegerField(default=0) # stored as a positive number
+    buildings_upkeep = models.BigIntegerField(default=0) # stored as a positive number
+    units_upkeep = models.BigIntegerField(default=0) # stored as a positive number
     portals_upkeep = models.IntegerField(default=0) # stored as a positive number
-    population_upkeep_reduction = models.IntegerField(default=0)
+    population_upkeep_reduction = models.BigIntegerField(default=0)
 
     total_solar_collectors = models.IntegerField(default=staring_solars) # all of these are across all planets (which is why its in status and not planet)
     total_fission_reactors = models.IntegerField(default=0)
